[se_aracaju] Adiciona controle por data
Reorganiza código para seguir fluxo de leitura

Renomeia métodos

Move se_aracaju para diretório /se

wip
